chore: exports PageProvider types and functions
