+++ packages/core/src/typings/pageProvider.ts
@@ -1,0 +1,1 @@
+export * from 'src/sdk/overrides/PageProvider'
